+++ dbc_influxdb/common.py
@@ -2,6 +2,7 @@ import pytz
 
 # Column names of columns that are used as tags
 tags = [
+    'site',
     'varname',
     'units',
     'raw_varname',
@@ -15,16 +16,18 @@ tags = [
     'filegroup',
     'config_filetype',
     'data_version',
-    'gain'
+    'gain',
+    'offset'
 ]
 
-def convert_ts_to_timezone(timezone_offset_to_utc_hours:int,
+
+def convert_ts_to_timezone(timezone_offset_to_utc_hours: int,
                            timestamp_index):
     """Convert timestamp index to timezone
 
     Convert 'TIMESTAMP_END' to desired timezone, e.g. to CET,
     using the pytz package. pytz is quite flexible with GMT and fixed offsets,
-    and GMT is the same as UTC (no offset to UTC).
+    and here(?) GMT is the same as UTC (no offset to UTC).
 
     From: https://pvlib-python.readthedocs.io/en/v0.3.0/timetimezones.html#fixed-offsets:
       "The 'Etc/GMT*' time zones mentioned above provide fixed offset